requirements.txt
asgiref==3.3.1
pytz==2020.5
sqlparse==0.4.1
requests
psycopg2-binary
Django==3.1.5
djangorestframework==3.12.2
django-jsonfield
djoser
